clean up packet byte array exports
