+++ core/src/packets/ethernet.rs
@@ -447,58 +447,10 @@ impl SizeOf for EthernetHeader {
     }
 }
 
-/// VLAN 802.1Q byte array.
-#[cfg(any(test, feature = "testils"))]
-#[cfg_attr(docsrs, doc(cfg(feature = "testils")))]
-#[rustfmt::skip]
-pub const VLAN_802_1Q_PACKET: [u8; 64] = [
-// Ethernet header
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x01,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x02,
-    // tpid
-    0x81, 0x00,
-    // tci
-    0x00, 0x7b,
-    // ether type
-    0x08, 0x06,
-    // payload
-    0x00, 0x01, 0x08, 0x00, 0x06, 0x04, 0x00, 0x02, 0x00, 0x19,
-    0x06, 0xea, 0xb8, 0xc1, 0xc0, 0xa8, 0x7b, 0x01, 0xff, 0xff,
-    0xff, 0xff, 0xff, 0xff, 0xc0, 0xa8, 0x7b, 0x01, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-];
-
-/// VLAN 802.1AD byte array.
-#[cfg(any(test, feature = "testils"))]
-#[cfg_attr(docsrs, doc(cfg(feature = "testils")))]
-#[rustfmt::skip]
-pub const VLAN_802_1AD_PACKET: [u8; 68] = [
-// Ethernet header
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x01,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x02,
-    // tpid
-    0x88, 0xa8,
-    // tci
-    0x00, 0x1e,
-    // tpid
-    0x81, 0x00,
-    // tci
-    0x20, 0x65,
-    // ether type
-    0x08, 0x06,
-    // payload
-    0x00, 0x01, 0x08, 0x00, 0x06, 0x04, 0x00, 0x02, 0x00, 0x19,
-    0x06, 0xea, 0xb8, 0xc1, 0xc0, 0xa8, 0x7b, 0x01, 0xff, 0xff,
-    0xff, 0xff, 0xff, 0xff, 0xc0, 0xa8, 0x7b, 0x01, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x00
-];
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::packets::UDP_PACKET;
+    use crate::testils::byte_arrays::{IPV4_UDP_PACKET, VLAN_802_1AD_PACKET, VLAN_802_1Q_PACKET};
 
     #[test]
     fn size_of_ethernet_header() {
@@ -515,7 +467,7 @@ mod tests {
 
     #[capsule::test]
     fn parse_ethernet_packet() {
-        let packet = Mbuf::from_bytes(&UDP_PACKET).unwrap();
+        let packet = Mbuf::from_bytes(&IPV4_UDP_PACKET).unwrap();
         let ethernet = packet.parse::<Ethernet>().unwrap();
 
         assert_eq!("00:00:00:00:00:01", ethernet.dst().to_string());
@@ -549,7 +501,7 @@ mod tests {
 
     #[capsule::test]
     fn swap_addresses() {
-        let packet = Mbuf::from_bytes(&UDP_PACKET).unwrap();
+        let packet = Mbuf::from_bytes(&IPV4_UDP_PACKET).unwrap();
         let mut ethernet = packet.parse::<Ethernet>().unwrap();
         ethernet.swap_addresses();
