+++ core/src/packets/icmp/v4/mod.rs
@@ -371,37 +371,12 @@ impl Icmpv4Parse for Ipv4 {
     }
 }
 
-/// ICMPv4 packet as byte-array.
-#[cfg(any(test, feature = "testils"))]
-#[cfg_attr(docsrs, doc(cfg(feature = "testils")))]
-#[rustfmt::skip]
-pub const ICMPV4_PACKET: [u8; 74] = [
-// Ethernet header
-    0x00, 0x50, 0x56, 0xe0, 0x14, 0x49,
-    0x00, 0x0c, 0x29, 0x34, 0x0B, 0xde,
-    0x08, 0x00,
-// IPv4 header
-    0x45, 0x00, 0x00, 0x3c,
-    0xd7, 0x43, 0x00, 0x00,
-    0x80, 0x01, 0x2b, 0x73,
-    0xc0, 0xa8, 0x9e, 0x8b,
-    0xae, 0x89, 0x2a, 0x4d,
-// ICMPv4 header
-    0x08, 0x00, 0x2a, 0x5c, 0x02, 0x00,
-    0x21, 0x00, 0x61, 0x62, 0x63, 0x64,
-    0x65, 0x66, 0x67, 0x68, 0x69, 0x6a,
-    0x6b, 0x6c, 0x6d, 0x6e, 0x6f, 0x70,
-    0x71, 0x72, 0x73, 0x74, 0x75, 0x76,
-    0x77, 0x61, 0x62, 0x63, 0x64, 0x65,
-    0x66, 0x67, 0x68, 0x69,
-];
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::packets::ip::v4::Ipv4;
     use crate::packets::Ethernet;
-    use crate::testils::byte_arrays::UDP_PACKET;
+    use crate::testils::byte_arrays::{ICMPV4_PACKET, IPV4_UDP_PACKET};
     use crate::Mbuf;
 
     #[test]
@@ -423,7 +398,7 @@ mod tests {
 
     #[capsule::test]
     fn parse_non_icmpv4_packet() {
-        let packet = Mbuf::from_bytes(&UDP_PACKET).unwrap();
+        let packet = Mbuf::from_bytes(&IPV4_UDP_PACKET).unwrap();
         let ethernet = packet.parse::<Ethernet>().unwrap();
         let ipv4 = ethernet.parse::<Ipv4>().unwrap();
