+++ core/src/packets/icmp/v6/ndp/router_advert.rs
@@ -265,53 +265,6 @@ impl<E: Ipv6Packet> fmt::Debug for Icmpv6<E, RouterAdvertisement> {
     }
 }
 
-/// ICMPv6 Router Advertisement packet as byte-array.
-#[cfg(any(test, feature = "testils"))]
-#[cfg_attr(docsrs, doc(cfg(feature = "testils")))]
-#[rustfmt::skip]
-pub const ROUTER_ADVERT_PACKET: [u8; 142] = [
-// ethernet header
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x01,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x02,
-    0x86, 0xDD,
-// IPv6 header
-    0x60, 0x00, 0x00, 0x00,
-    // payload length
-    0x00, 0x58,
-    0x3a,
-    0xff,
-    0xfe, 0x80, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xd4, 0xf0, 0x45, 0xff, 0xfe, 0x0c, 0x66, 0x4b,
-    0xff, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01,
-// ICMPv6 header
-    // type
-    0x86,
-    // code
-    0x00,
-    // checksum
-    0xf5, 0x0c,
-// router advertisement message
-    // current hop limit
-    0x40,
-    // flags
-    0x40,
-    // router lifetime
-    0x0e, 0x10,
-    // reachable time
-    0x00, 0x00, 0x00, 0x00,
-    // retrans timer
-    0x00, 0x00, 0x00, 0x00,
-    // prefix information option
-    0x03, 0x04, 0x40, 0xc0, 0x00, 0x00, 0x09, 0x3e, 0x00, 0x00, 0x09, 0x3e, 0x00, 0x00, 0x00, 0x00,
-    0x26, 0x07, 0xfc, 0xc8, 0xf1, 0x42, 0xb0, 0xf0, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-    // MTU option
-    0x05, 0x01, 0x00, 0x00, 0x00, 0x00, 0x05, 0xdc,
-    // source link-layer address option
-    0x01, 0x01, 0x70, 0x3a, 0xcb, 0x1b, 0xf9, 0x7a,
-    // recursive DNS server option
-    0x19, 0x03, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff, 0x26, 0x07, 0xfc, 0xc8, 0xf1, 0x42, 0xb0, 0xf0,
-    0xd4, 0xf0, 0x45, 0xff, 0xfe, 0x0c, 0x66, 0x4b
-];
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -322,6 +275,7 @@ mod tests {
     use crate::packets::icmp::v6::{Icmpv6Message, Icmpv6Parse, Icmpv6Types};
     use crate::packets::ip::v6::Ipv6;
     use crate::packets::{Ethernet, Packet};
+    use crate::testils::byte_arrays::ROUTER_ADVERT_PACKET;
     use crate::{Mbuf, SizeOf};
     use fallible_iterator::FallibleIterator;
     use std::str::FromStr;
