+++ core/src/packets/icmp/v6/ndp/router_solicit.rs
@@ -91,43 +91,13 @@ impl<E: Ipv6Packet> fmt::Debug for Icmpv6<E, RouterSolicitation> {
     }
 }
 
-/// ICMPv6 Router Solicitation packet as byte-array.
-#[cfg(any(test, feature = "testils"))]
-#[cfg_attr(docsrs, doc(cfg(feature = "testils")))]
-#[rustfmt::skip]
-pub const ROUTER_SOLICIT_PACKET: [u8; 70] = [
-    // ** ethernet header
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x01,
-    0x00, 0x00, 0x00, 0x00, 0x00, 0x02,
-    0x86, 0xDD,
-    // ** IPv6 header
-    0x60, 0x00, 0x00, 0x00,
-    // payload length
-    0x00, 0x10,
-    0x3a,
-    0xff,
-    0xfe, 0x80, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xd4, 0xf0, 0x45, 0xff, 0xfe, 0x0c, 0x66, 0x4b,
-    0xff, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01,
-    // ** ICMPv6 header
-    // type
-    0x85,
-    // code
-    0x00,
-    // checksum
-    0xf5, 0x0c,
-    // ** router solicitation message
-    // reserved
-    0x00, 0x00, 0x00, 0x00,
-    // ** source link-layer address option
-    0x01, 0x01, 0x70, 0x3a, 0xcb, 0x1b, 0xf9, 0x7a
-];
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::packets::icmp::v6::{Icmpv6Message, Icmpv6Parse, Icmpv6Types};
     use crate::packets::ip::v6::Ipv6;
     use crate::packets::{Ethernet, Packet};
+    use crate::testils::byte_arrays::ROUTER_SOLICIT_PACKET;
     use crate::{Mbuf, SizeOf};
 
     #[test]
