+++ core/src/packets/ip/v4.rs
@@ -577,7 +577,7 @@ impl Header for Ipv4Header {}
 mod tests {
     use super::*;
     use crate::packets::ip::ProtocolNumbers;
-    use crate::testils::byte_arrays::{IPV6_PACKET, UDP_PACKET};
+    use crate::testils::byte_arrays::{IPV4_UDP_PACKET, IPV6_TCP_PACKET};
     use crate::Mbuf;
 
     #[test]
@@ -587,7 +587,7 @@ mod tests {
 
     #[capsule::test]
     fn parse_ipv4_packet() {
-        let packet = Mbuf::from_bytes(&UDP_PACKET).unwrap();
+        let packet = Mbuf::from_bytes(&IPV4_UDP_PACKET).unwrap();
         let ethernet = packet.parse::<Ethernet>().unwrap();
         let ipv4 = ethernet.parse::<Ipv4>().unwrap();
 
@@ -609,7 +609,7 @@ mod tests {
 
     #[capsule::test]
     fn parse_non_ipv4_packet() {
-        let packet = Mbuf::from_bytes(&IPV6_PACKET).unwrap();
+        let packet = Mbuf::from_bytes(&IPV6_TCP_PACKET).unwrap();
         let ethernet = packet.parse::<Ethernet>().unwrap();
 
         assert!(ethernet.parse::<Ipv4>().is_err());
@@ -617,7 +617,7 @@ mod tests {
 
     #[capsule::test]
     fn parse_ipv4_setter_checks() {
-        let packet = Mbuf::from_bytes(&UDP_PACKET).unwrap();
+        let packet = Mbuf::from_bytes(&IPV4_UDP_PACKET).unwrap();
         let ethernet = packet.parse::<Ethernet>().unwrap();
         let mut ipv4 = ethernet.parse::<Ipv4>().unwrap();
 
@@ -675,7 +675,7 @@ mod tests {
 
     #[capsule::test]
     fn compute_checksum() {
-        let packet = Mbuf::from_bytes(&UDP_PACKET).unwrap();
+        let packet = Mbuf::from_bytes(&IPV4_UDP_PACKET).unwrap();
         let ethernet = packet.parse::<Ethernet>().unwrap();
         let mut ipv4 = ethernet.parse::<Ipv4>().unwrap();
